+++ pesdk-v5/types/components/base/cardLabel.d.ts
@@ -1,0 +1,12 @@
+/// <reference types="react" />
+export interface CustomCardLabelProps {
+    className?: string;
+    caret?: boolean;
+    isActive?: boolean;
+    withBackground?: boolean;
+    children: React.ReactNode;
+    style?: {
+        [key: string]: string;
+    };
+    hidden?: boolean;
+}