+++ pesdk-v5/types/components/base/checkbox/checkbox.d.ts
@@ -1,0 +1,14 @@
+/// <reference types="react" />
+export interface CustomCheckboxProps {
+    label: string;
+    value: boolean;
+    isDisabled?: boolean;
+    onClick: (e?: React.SyntheticEvent) => void;
+    checkMarkPosition?: 'left' | 'right';
+    className?: string;
+    style?: {
+        [key: string]: string;
+    };
+    children?: React.ReactNode;
+}
+export declare type CustomCheckbox = React.ComponentType<CustomCheckboxProps>;