+++ pesdk-v5/types/components/base/color/colorItemComponents.d.ts
@@ -1,0 +1,31 @@
+/// <reference types="react" />
+export interface ColorItemBaseProps {
+    onClick: (e: React.MouseEvent<HTMLButtonElement>) => void;
+    disabled: boolean;
+    className?: string;
+    style?: {
+        [key: string]: string;
+    };
+    children?: React.ReactNode;
+}
+export interface ColorItemBackgroundProps {
+    color: string;
+    className?: string;
+    style?: {
+        [key: string]: string;
+    };
+}
+export interface ColorItemTiledBackgroundProps {
+    url: string;
+    className?: string;
+    style?: {
+        [key: string]: string;
+    };
+}
+export interface ColorItemActiveOverlayProps {
+    isActive: boolean;
+    className?: string;
+    style?: {
+        [key: string]: string;
+    };
+}