+++ pesdk-v5/types/components/base/label.d.ts
@@ -1,0 +1,10 @@
+export interface CustomLabelProps {
+    label: string;
+    htmlFor?: string;
+    isDisabled?: boolean;
+    show?: boolean;
+    className?: string;
+    style?: {
+        [key: string]: string;
+    };
+}