+++ pesdk-v5/types/components/base/loader.d.ts
@@ -1,0 +1,12 @@
+/// <reference types="react" />
+export interface CustomLoaderProps {
+    show: boolean;
+    heading: string;
+    body: string;
+    position?: {
+        x: number;
+        y: number;
+    };
+    className?: string;
+}
+export declare type CustomLoader = React.ComponentType<CustomLoaderProps>;