+++ pesdk-v5/types/configuration/custom/loader.d.ts
@@ -1,5 +1,0 @@
-/// <reference types="react" />
-export interface CustomLoaderProps {
-    show: boolean;
-}
-export declare type CustomLoader = React.ElementType<CustomLoaderProps>;