+++ pesdk-v5/types/configuration/custom/toolbarItem.d.ts
@@ -1,9 +1,0 @@
-/// <reference types="react" />
-export interface CustomToolbarItemProps {
-    tool: string;
-    label: string;
-    icon: React.ReactNode;
-    isActive: boolean;
-    onClick: (e?: any) => void;
-}
-export declare type CustomAdvancedToolbarItem = React.ElementType<CustomToolbarItemProps>;